[sensorfw] Activate sensors on sensorfwd restart. Fixes JB#41999
Sometimes lipstick does not react to orientation changes and sensorfwd
restarts do not help. Looks the fault is triggered when sensorfwd is not
running or orientation sensor is not available during lipstick startup.
The reason is that m_attemptRestart flag is set only when a start attempt
fails due to sensorfwd not being present - within a conditional block that
is not entered until sensorfwd has been seen to be available. Which then
means that any initial start failure can get locked in, and no retries are
made by qtsensors code within the application's life time.

Rename m_attemptRestart to m_startRequested. Set/clear it unconditionally
on start()/stop() call. When sensorfwd startup is detected, make an
attempt to start flagged sensors.

Also drop the practically unused "running" member variable.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>